correction poids miel en kg double

+++ src/main/java/ooioo/ruches/recolte/RecolteRepository.java
@@ -36,7 +36,7 @@ public interface RecolteRepository extends CrudRepository<Recolte, Long> {
 	  from recolte 
 	  where date_part('year', date) = ?1
 			""", nativeQuery = true)
-	Optional<Integer> findPoidsMielByYear(Double year);
+	Optional<Double> findPoidsMielByYear(Double year);
 
 	/*
 	 * Renvoie le poids de miel récolté dans les hausses pour l'année passée en paramètre
